+++ primes.h
@@ -1,4 +1,4 @@
-bool isPrime (int n);
-bool isPrimeMem (int n, int* primes);
-int listOfPrimes (int n, int** primes_ptr, bool** primesTable_ptr);
+bool isPrime (long n);
+bool isPrimeMem (long n, long* primes);
+long listOfPrimes (long n, long** primes_ptr, bool** primesTable_ptr);
 void sieveOfEratosthenes (int n, bool** primesTable_ptr);
